data science/back/requirements.txt
# --- Web 框架 ---
Flask==3.0.0
gunicorn==21.2.0
flask-cors==4.0.0

# --- Google Cloud & Firebase ---
firebase-admin==6.5.0
google-cloud-storage==2.14.0

# --- 环境变量管理 ---
python-dotenv==1.0.0

# --- 测试工具 ---
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# --- 数据科学 ---
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4  # 统计检验
openpyxl==3.1.2  # Excel文件支持
# scikit-learn==1.3.2  # 需要时取消注释
# matplotlib==3.8.2
# seaborn==0.13.0